Move PathBuilder to own file
